Migrating JCPJSRLocator to better design
* Also, make it use the HttpService
